+++ modules/hanna-react/src/DropdownButton.tsx
@@ -1,7 +1,15 @@
-import React, { CSSProperties, MutableRefObject, useState } from 'react';
+import React, {
+  CSSProperties,
+  HTMLAttributeAnchorTarget,
+  MouseEvent,
+  MutableRefObject,
+  ReactElement,
+  useRef,
+  useState,
+} from 'react';
 import { autoUpdate, flip, shift, useFloating } from '@floating-ui/react';
 import { IconName } from '@reykjavik/hanna-css';
-import { modifiedClass } from '@reykjavik/hanna-utils';
+import { ClassNameModifiers, EitherObj, modifiedClass } from '@reykjavik/hanna-utils';
 
 import { Button, ButtonVariantProps } from './_abstract/_Button.js';
 import { useCallbackOnEsc } from './utils/useCallbackOnEsc.js';
@@ -17,29 +25,80 @@ type Prefix<record extends Record<string, unknown>, prefix extends string> = {
 
 // ---------------------------------------------------------------------------
 
-export type DropdownButtonItem = MainMenu2Item & { icon?: IconName };
+export type DropdownButtonItem = {
+  /** Visible label text */
+  label: string | ReactElement;
+  /** Un-abbreviated label set as `title=""` and `aria-label=""` */
+  labelLong?: string;
+  /** Language of the link label */
+  lang?: string;
+  /** Languge of the linked resource */
+  hrefLang?: string;
 
-export type DropdownButtonCustomItem = (props: {
-  closeMenu: () => void;
-}) => React.ReactElement;
+  /**
+   * Puts a modifier className for the menu __item <li/> element.
+   * */
+  modifier?: ClassNameModifiers;
+  /** Signifies if the menu item is part of the page's breadcrumb trail */
+  current?: boolean;
+
+  /**
+   * The URL the link points to.
+   *
+   * If neither `href` nor `onClick` is passed, then the item is not rendered
+   * at all.
+   */
+  href?: string;
+  /** Sets `target=""` on anchor tags with a `href` attribute. */
+  target?: HTMLAttributeAnchorTarget;
+
+  /**
+   * Adding `onClick` automatically results in a <button/> element being
+   * rendered. If `href` is also passed, then a <a href/> element is rendered
+   * during initial (server-side) render, which then gets replaced by a
+   * <button/> element during the first client-side
+   *
+   * NOTE: Clicking a menu item will automatically close tghe menu
+   * … unless the `onClick` function explicitly returns `false`.
+   */
+  onClick?: (item: MainMenu2Item) => void | boolean;
+  /** Sets `aria-controls=""` on `<button/>`s with `onClick` */
+  controlsId?: string;
+
+  Content?: never; // To discrimiate bteween this and MainMenu2CustomItem
+
+  icon?: IconName;
+};
+
+type DropdownButtonCustomItemFn = (props: { closeMenu: () => void }) => ReactElement;
+
+export type DropdownButtonCustomItem = Pick<
+  DropdownButtonItem,
+  'modifier' | 'current'
+> & {
+  Content: DropdownButtonCustomItemFn;
+};
 
 //
 
 export type DropdownButtonProps = {
-  label: string | NonNullable<React.ReactElement>;
-  labelLong?: string;
   items: Array<DropdownButtonItem | DropdownButtonCustomItem>;
-
   /**
-   * NOTE: Clicking a MainMenu2 item will automatically close HannaUIState's
-   * "Hamburger menu" (a.k.a. "Mobile menu")
+   * NOTE: Clicking a DropdownButton item will automatically close the drropdown
    * … unless the `onItemClick` function explicitly returns `false`.
    */
   onItemClick?: (item: MainMenu2Item) => void | boolean;
-
-  /** Default: `"seconcary"` */
-  buttonType?: 'primary' | 'secondary';
-} & Prefix<Omit<ButtonVariantProps, 'small'>, 'button'> &
+} & EitherObj<
+  {
+    label: string | ReactElement;
+    labelLong?: string;
+    /** Default: `"seconcary"` */
+    buttonType?: 'primary' | 'secondary';
+  } & Prefix<Omit<ButtonVariantProps, 'small'>, 'button'>,
+  {
+    Toggler: (props: { isOpen: boolean }) => ReactElement;
+  }
+> &
   WrapperElmProps<'details', 'open' | 'name'> &
   SSRSupportProps;
 
@@ -47,7 +106,7 @@ export const DropdownButton = (props: DropdownButtonProps) => {
   const [isOpen, setIsOpen] = useLaggedState(false, 10);
   const isBrowser = useIsBrowserSide(props.ssr);
   const [isHovering, setIsHovering] = useState(false);
-  const wrapperRef = React.useRef<HTMLDetailsElement>(null);
+  const wrapperRef = useRef<HTMLDetailsElement>(null);
 
   const closeMenuStat = () => setIsOpen(false, 0);
 
@@ -61,6 +120,11 @@ export const DropdownButton = (props: DropdownButtonProps) => {
   });
 
   const { onItemClick, wrapperProps = {} } = props;
+
+  const toggle = (e: MouseEvent) => {
+    e.preventDefault();
+    setIsOpen(!isOpen, 0);
+  };
 
   return (
     <details
@@ -86,22 +150,24 @@ export const DropdownButton = (props: DropdownButtonProps) => {
         refs.setFloating(elm.querySelector('.DropdownButton__menu'));
       }}
     >
-      <Button
-        as="summary"
-        className="DropdownButton__toggler"
-        bem={props.buttonType === 'primary' ? 'ButtonPrimary' : 'ButtonSecondary'}
-        icon={props.buttonIcon}
-        size={props.buttonSize}
-        variant={props.buttonVariant}
-        aria-label={props.labelLong}
-        onClick={(e) => {
-          e.preventDefault();
-          setIsOpen(!isOpen, 0);
-        }}
-      >
-        {props.label}
-      </Button>
-
+      {props.Toggler ? (
+        <summary className="DropdownButton__toggler" onClick={toggle}>
+          <props.Toggler isOpen={isOpen} />
+        </summary>
+      ) : (
+        <Button
+          as="summary"
+          className="DropdownButton__toggler"
+          bem={props.buttonType === 'primary' ? 'ButtonPrimary' : 'ButtonSecondary'}
+          icon={props.buttonIcon}
+          size={props.buttonSize}
+          variant={props.buttonVariant}
+          aria-label={props.labelLong}
+          onClick={toggle}
+        >
+          {props.label}
+        </Button>
+      )}
       <ul
         className="DropdownButton__menu"
         onMouseEnter={() => {
@@ -124,13 +190,22 @@ export const DropdownButton = (props: DropdownButtonProps) => {
       >
         {props.items.map((item, i) => {
           if (typeof item === 'function') {
-            const Item = item;
+            item = { Content: item };
+          }
+
+          const itemProps = {
+            className: modifiedClass('DropdownButton__item', item.modifier),
+            'aria-current': item.current || undefined,
+          };
+
+          if ('Content' in item && item.Content) {
             return (
-              <li key={i} className="DropdownButton__item">
-                <Item closeMenu={closeMenuStat} />
+              <li key={i} data-customitem="" {...itemProps}>
+                <item.Content closeMenu={closeMenuStat} />
               </li>
             );
           }
+
           const { label, onClick, href } = item;
 
           const commonProps = {
@@ -141,6 +216,10 @@ export const DropdownButton = (props: DropdownButtonProps) => {
           };
 
           const doRenderButton = isBrowser && (onClick || (onItemClick && href == null));
+
+          // TypeScript type-narrowing helper for the onClick callbacks below — because
+          // `item` is a variable and could hypothetically change before the click occurs
+          const _item = item;
 
           return (
             <li
@@ -154,8 +233,8 @@ export const DropdownButton = (props: DropdownButtonProps) => {
                   type="button"
                   aria-controls={item.controlsId}
                   onClick={() => {
-                    const keepOpen1 = onClick && onClick(item) === false;
-                    const keepOpen2 = onItemClick && onItemClick(item) === false;
+                    const keepOpen1 = onClick && onClick(_item) === false;
+                    const keepOpen2 = onItemClick && onItemClick(_item) === false;
                     !(keepOpen1 || keepOpen2) && closeMenuStat();
                   }}
                 >
@@ -168,7 +247,7 @@ export const DropdownButton = (props: DropdownButtonProps) => {
                   hrefLang={item.hrefLang}
                   target={item.target}
                   onClick={() => {
-                    const keepOpen = onItemClick && onItemClick(item) === false;
+                    const keepOpen = onItemClick && onItemClick(_item) === false;
                     !keepOpen && closeMenuStat();
                   }}
                 >